refactor(web): align transfer dialog file name
