+++ web/src/features/projects/components/TransferProjectDialogController.tsx
@@ -1,7 +1,7 @@
 import { type ReactNode, useState } from "react";
 
 import { Dialog } from "@/src/components/ui/dialog";
-import { TransferProjectDialogContent } from "@/src/features/projects/components/TransferProjectDialog";
+import { TransferProjectDialogContent } from "@/src/features/projects/components/TransferProjectDialogContent";
 import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
 import {
   hasOrganizationAccess,
